+++ src/ui/EditPopup.cpp
@@ -33,7 +33,7 @@ bool EditPopup::setup(ObjectData obj, std::unordered_set<std::string> availableT
             m_objDesc->getInputNode()->m_placeholderLabel->setOpacity((p0.length() == 0) ? 255 : 0);
             textArea->setScale(Utils::calculateScale(p0, 50, 300, 1.0F, 0.35F));
             textArea->m_width = 220.0F / Utils::calculateScale(p0, 50, 300, 1.0F, 0.32F);
-            textArea->setString(p0);
+            textArea->setString(p0.data());
         }
     );